+++ src/RateLimit/NoRateLimit.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tomaj\NetteApi\RateLimit;
+
+class NoRateLimit implements RateLimitInterface
+{
+    public function check(): ?RateLimitResponse
+    {
+        return null;
+    }
+}