+++ src/RateLimit/RateLimitInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tomaj\NetteApi\RateLimit;
+
+interface RateLimitInterface
+{
+    public function check(): ?RateLimitResponse;
+}